requirements.txt
fastapi
uvicorn==0.20.0
pymongo==4.3.3
redis==4.5.1
pandas==1.3.3
starlette==0.25.0
requests==2.28.2
Pillow==9.4.0
numpy==1.19.5
scikit-learn==1.2.1
tensorflow
pydot
graphviz
typing-extensions